Always install the ceph apache deb repo
When the cookbook is evaluated, the ceph::repo recipe is run before the
ceph::radosgw recipe, and so the check to include the apache repo
always fails. This changes it to always include it if we have the
node['ceph']['radosgw']['use_apache_fork'] attribute is true.

Debian only has the Apache fastcgi plugin in the non-free component,
which isn't enabled in the Vagrant testkitchen image. Ubuntu has it
available in the default repo.