Automatically generates a monsecret if it is blank
Some of the more advanced features require a mon secret to
generate auth keys for other things, such as radosgw.
This change will run gen-key command to generate a mon secret if it
has not been set as an attribute already. It will also search for
other mon nodes and use the cluster secret if it already exists.